requirements.txt
flask==3.0.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
paho-mqtt==2.1.0
gunicorn==22.0.0